03_ai_protocols/01_foundation_protocols/06_gRPC/grpc-tutorial/pyproject.toml
[project]
name = "grpc-tutorial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "grpcio>=1.71.0",
    "grpcio-tools>=1.71.0",
]